Refactor Years Pane to its own module
