+++ src/components/Visualizations/index.tsx
@@ -1,3 +1,4 @@
 export * from './MetricsPageContainer';
 export * from './OverviewPageContainer';
 export * from './VisualizationTabs';
+export * from './YearsPane';
